requirements.txt

gunicorn
django-heroku

Django==2.2.1

djangorestframework==3.10.0

dynamic-rest==1.9.4

factory-boy==2.11.1

oauth==1.0.1


